Made the things work (mostly)
So the code itself now works to do what it's supposed to, though it's still not where it needs to be in terms of actually being implemented. There's lots of stupid hard coding too (Amber made me do it.) However, something in Robot.java is preventing it from working properly and seems to be setting the setpoint to 0. This needs to be fixed before this branch is merged in.

NOTHING IS CALIBRATED RIGHT NOW. DO NOT MERGE YET.