Started on selection box.
Wow this was annoying.
Polygon lacks the ability to do much of anything with verticies, so the custom class BoxPanel will be doing the interesting stuff then translating that to a Polygon for painting.
Thanks to the fact that WebcamPanel gets black bars injected all over but needs to remain in sync regarding size with BoxPanel, a GridBagLayout wrapper needs to be used to maintain all the sizing. GridBagLayout is hard.